Firewall | Error reporting status
The error state is because the firewall setting is not ready immediately after the machine restarts or wakes up from sleep. If the device checks in immediately, the compliance status of the firewall setting will return as "Error". In this case, this error state is by design and won't affect device's overall compliance state immediately.