Update device-restrictions-macos.md
The "Device enrollment" headers also apply to devices enrolled with automated device enrollment (formerly DEP). All settings that apply to device enrolled devices can also apply to automated device enrolled devices. This was previously missed. The UI will mirror this with the 1911 release, but we should have the documentation be correct as soon as possible.