Update device-restrictions-ios.md
Plan for Change: The ‘Server-side Logging for Siri commands’ setting will be removed from the Intune console
We plan to remove the setting “Server-side logging for Siri commands” from the Intune console with the November update to the Intune service. It currently isn’t working on devices even though you can see it in the management profile if a device is targeted, because Apple has removed it.

How does this affect me? :
When the November update or 1911 rolls out around mid-November, you’ll see that this setting has been removed from the Device restrictions menu (Built-in Apps) for iOS configuration profiles. Apple has already removed this setting so Intune is removing it as well.  It may appear in your policies and the targeted device’s management profile until your policy is recomputed, but the setting has no affect on your device. If you use this setting, know that if you go back to the profile with this setting, and you make a minor edit and save the policy, the policy will recompute and the setting will be deleted from your policy. We do not anticipate any impact since it is unsupported by Apple. 
If you choose not to make a minor edit to the policies that have this setting configured, end users will see this setting in the management profile of their device but the setting has no effect. 

What can I do to prepare for this change?:
If you want this setting deleted from your profiles, go to the profile with the setting configured, make a minor change, and save the profile. The policy will recomute and the setting will be deled from targeted devices and the policy. If you don’t care whether the setting is in the policy or on the device, you can leave everything as is. We’ll update our What’s New page and documentation when this change rolls out.