SCEP Certificate + ESP issue is fixed
It is not only documented by Michael Niehaus, I also tested it successful in a lab environment. Issue can be treated as resolved!

Windows Autopilot and ESP now work with Windows Hello for Business certs
https://oofhours.com/2019/10/29/windows-autopilot-and-esp-now-work-with-windows-hello-for-business-certs/

So, we implemented a change in Intune that will work around this.  Instead of delivering the certificate policies during user ESP, it will instead wait until later, allowing ESP to complete and the WHfB enrollment process to complete.  So if you had disabled user ESP because of this issue, try turning it back on again.