Clarified opening text regarding user interaction fff
We noted that no user interaction is required to deploy the update - however this is incorrect. Modified this to align with the Apple documentation on this: https://developer.apple.com/business/documentation/MDM-Protocol-Reference.pdf (page 76 specifically).
"On iOS 10.3 and later, supported Software Update commands require supervision but not DEP enrollment. If there is
a passcode on the device, a user must enter it to start a software update."

Not sure if we need to cross-reference their document too?