Add a warning about notarization for macOS LOB apps
LOB apps that have not been notarized will fail to run starting with Catalina. Customers need to check if their lob apps are notarized before using the app packaging tool to add the apps to Intune. The app wrapping tool does not have a notarization check, so the customers must check.