Remove incorrect note about app config and APP
Revert #2439. I did a bit of digging into why the docs say app protection policy is required for app configuration policy. This seems to be an outdated note that only applied to the Managed Browser when it did not have an enrollment flow of its own and relied on auto-enrollment. Whatever the reason, it is incorrect. I tested out signing in to Edge with a user targeted with config policy but with no protection policy and I confirmed that the config policy does get received.