Update identity-protection-configure.md
Proposing to add following note based on the customers support experience:
Important:  If you need to allow multiple users to be provisioned to a device, you must specify that the Windows Hello for Business policy be applied to the devices. If the policy is only applied to the users, then only one user can be provisioned to a device.